Ajout des exceptions pour les operations crud des 3 entites

+++ src/main/java/com/safetyNet/service/MedicalRecordsService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.safetyNet.exceptions.MedicalRecordsIntrouvableException;
 import com.safetyNet.model.MedicalRecordsModel;
 
 import com.safetyNet.repository.MedicalRecordRepository;
@@ -19,12 +20,14 @@ public class MedicalRecordsService {
 	MedicalRecordRepository medicalRecordRepository;
 
 	// Recuperer tous les MedicalRecords
-	public List<MedicalRecordsModel> getMedicalRecords() {
+	public List<MedicalRecordsModel> getMedicalRecords() throws MedicalRecordsIntrouvableException {
+		if(medicalRecordRepository.findAll().isEmpty()) throw new MedicalRecordsIntrouvableException("La liste des dossiers est vide");
 		return medicalRecordRepository.findAll();
 	}
 
 	// Recuperer un MedicalRecord
-	public MedicalRecordsModel getMedicalRecord(String firstName, String lastName) {
+	public MedicalRecordsModel getMedicalRecord(String firstName, String lastName) throws MedicalRecordsIntrouvableException {
+		if(medicalRecordRepository.findByfirstName(firstName, lastName) ==  null )throw  new MedicalRecordsIntrouvableException("Le dossier de "+ firstName + "  " + lastName + " est introuvable");
 		return medicalRecordRepository.findByfirstName(firstName, lastName);
 	}
 
@@ -34,6 +37,7 @@ public class MedicalRecordsService {
 
 		if (newMedicalRecord != null) {
 			medicalRecordRepository.saveMedicalRecord(newMedicalRecord);
+			
 			return ResponseEntity.ok("le dossier est ajouté avec succès.");
 		} else {
 			return ResponseEntity.notFound().build();
@@ -41,10 +45,11 @@ public class MedicalRecordsService {
 
 	}
 
-	public ResponseEntity<String> updateMedicalRecord(MedicalRecordsModel updateMedicalRecord) {
+	public ResponseEntity<String> updateMedicalRecord(MedicalRecordsModel updateMedicalRecord) throws MedicalRecordsIntrouvableException {
 
 		if (updateMedicalRecord != null) {
 			medicalRecordRepository.updateMedicalRecord(updateMedicalRecord);
+			if(medicalRecordRepository.updateMedicalRecord(updateMedicalRecord) ==  null )throw new MedicalRecordsIntrouvableException("Le dossier de "+ updateMedicalRecord.firstName + "  " + updateMedicalRecord.lastName + " est introuvable");
 			return ResponseEntity.ok("le dossier est mis à jour avec succès.");
 		} else {
 			return ResponseEntity.notFound().build();
@@ -52,10 +57,12 @@ public class MedicalRecordsService {
 	}
 
 	// supprimer une dossier
-	public ResponseEntity<String> removeMedicalRecord(String firstName, String lastName) {
+	public ResponseEntity<String> removeMedicalRecord(String firstName, String lastName) throws MedicalRecordsIntrouvableException {
 
 		if (firstName != null && lastName != null) {
 			MedicalRecordsModel medicalRecordToDelet = medicalRecordRepository.findByfirstName(firstName, lastName);
+
+			if(medicalRecordToDelet ==  null )throw new MedicalRecordsIntrouvableException("Le dossier de "+ firstName + "  " + lastName + " est introuvable");
 			medicalRecordRepository.deleteMedicalRecordsModel(medicalRecordToDelet);
 			return ResponseEntity.ok("le dossier est supprimée avec succès.");
 		} else {
