+++ src/main/java/com/safetyNet/service/PersonsService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.safetyNet.exceptions.PersonIntrovableExeption;
 import com.safetyNet.model.PersonsModel;
 import com.safetyNet.repository.PersonsRepository;
 
@@ -17,11 +18,13 @@ public class PersonsService {
 	@Autowired
 	PersonsRepository personsRepository;
 
-	public List<PersonsModel> getPersons() {
+	public List<PersonsModel> getPersons() throws PersonIntrovableExeption {
+		if(personsRepository.findAll().isEmpty()) throw new PersonIntrovableExeption("La liste des personnes est vide");
 		return personsRepository.findAll();
 	}
 
-	public PersonsModel getPerson(String firstName, String lastName) {
+	public PersonsModel getPerson(String firstName, String lastName) throws PersonIntrovableExeption {
+		if(personsRepository.findByfirstName(firstName, lastName) == null) throw new PersonIntrovableExeption("La personne " +firstName+ " "+ lastName + " est introuvable");
 		return personsRepository.findByfirstName(firstName, lastName);
 	}
 
@@ -29,8 +32,7 @@ public class PersonsService {
 	public ResponseEntity<String> addPerson(PersonsModel person) {
 		
 		if(person !=  null)
-    	{
-			
+    	{	
 		   personsRepository.save(person);
      	   return ResponseEntity.ok("la personne est ajoutée avec succès.");
     	}
@@ -41,11 +43,12 @@ public class PersonsService {
 	}
 
 	// supprimer une personne 
-	public ResponseEntity<String>  removePerson(String firstName, String lastName) {
+	public ResponseEntity<String> removePerson(String firstName, String lastName) throws PersonIntrovableExeption {
 
 		if(firstName != null && lastName != null)
     	{
 			PersonsModel personToMDelet = personsRepository.findByfirstName(firstName, lastName);
+			if(personToMDelet == null) throw new PersonIntrovableExeption("La personne " +firstName+ " "+ lastName + " est introuvable");
 			personsRepository.deletePerson(personToMDelet);
 			return ResponseEntity.ok("la personne est supprimée avec succès.");
     	}
@@ -56,12 +59,13 @@ public class PersonsService {
 	}
 
 	//modifier une personne
-	public ResponseEntity<String> updatePerson(PersonsModel  newPreson) 
+	public ResponseEntity<String> updatePerson(PersonsModel  newPreson) throws PersonIntrovableExeption 
 	{
 		
 		if(newPreson != null )
     	{
 			personsRepository.updatePerson(newPreson);
+			if(personsRepository.updatePerson(newPreson) == null) throw new PersonIntrovableExeption("La personne " +newPreson.firstName+ " "+ newPreson.lastName + " est introuvable");
      	   return ResponseEntity.ok("Les informations de la personne sont mises à jour avec succès.");
     	}
 	    else
